+++ __tests__/factories/event.spec.ts
@@ -100,6 +100,30 @@ describe('event types', () => {
     expect(event.routing.target.name).toEqual(character.getId());
   });
 
+  test('should generate tts playback mute', () => {
+    const event = factory.ttsPlaybackMute(true);
+
+    expect(event).toHaveProperty('packetId');
+    expect(event).toHaveProperty('routing');
+    expect(event).toHaveProperty('timestamp');
+    expect(event.control).toEqual({
+      action: ControlEventAction.TTS_PLAYBACK_MUTE,
+    });
+    expect(event.routing.target.name).toEqual(character.getId());
+  });
+
+  test('should generate tts playback unmute', () => {
+    const event = factory.ttsPlaybackMute(false);
+
+    expect(event).toHaveProperty('packetId');
+    expect(event).toHaveProperty('routing');
+    expect(event).toHaveProperty('timestamp');
+    expect(event.control).toEqual({
+      action: ControlEventAction.TTS_PLAYBACK_UNMUTE,
+    });
+    expect(event.routing.target.name).toEqual(character.getId());
+  });
+
   test('should generate text event', () => {
     const text = v4();
     const event = factory.text(text);
@@ -111,7 +135,7 @@ describe('event types', () => {
     expect(event.routing.target.name).toEqual(character.getId());
   });
 
-  test('should generate trigger event', () => {
+  test('should generate trigger event without parameters', () => {
     const name = v4();
     const event = factory.trigger(name);
 
@@ -119,6 +143,20 @@ describe('event types', () => {
     expect(event).toHaveProperty('routing');
     expect(event).toHaveProperty('timestamp');
     expect(event.custom.name).toEqual(name);
+    expect(event.custom.parameters).toEqual(undefined);
+    expect(event.routing.target.name).toEqual(character.getId());
+  });
+
+  test('should generate trigger event with parameters', () => {
+    const name = v4();
+    const parameters = [{ name: v4(), value: v4() }];
+    const event = factory.trigger(name, parameters);
+
+    expect(event).toHaveProperty('packetId');
+    expect(event).toHaveProperty('routing');
+    expect(event).toHaveProperty('timestamp');
+    expect(event.custom.name).toEqual(name);
+    expect(event.custom.parameters).toEqual(parameters);
     expect(event.routing.target.name).toEqual(character.getId());
   });
 
@@ -178,21 +216,30 @@ describe('convert packet to external one', () => {
       timestamp: protoTimestamp(),
     };
 
-    const result = factory.convertToInworldPacket(packet);
+    const result = EventFactory.fromProto(packet);
 
     expect(result).toBeInstanceOf(InworldPacket);
     expect(result.isAudio()).toEqual(true);
   });
 
   test('text', () => {
-    const result = factory.convertToInworldPacket(factory.text(v4()));
+    const result = EventFactory.fromProto(factory.text(v4()));
 
     expect(result).toBeInstanceOf(InworldPacket);
     expect(result.isText()).toEqual(true);
   });
 
-  test('trigger', () => {
-    const result = factory.convertToInworldPacket(factory.trigger(v4()));
+  test('trigger without parameters', () => {
+    const result = EventFactory.fromProto(factory.trigger(v4()));
+
+    expect(result).toBeInstanceOf(InworldPacket);
+    expect(result.isTrigger()).toEqual(true);
+  });
+
+  test('trigger with parameters', () => {
+    const result = EventFactory.fromProto(
+      factory.trigger(v4(), [{ name: v4(), value: v4() }]),
+    );
 
     expect(result).toBeInstanceOf(InworldPacket);
     expect(result.isTrigger()).toEqual(true);
@@ -209,7 +256,7 @@ describe('convert packet to external one', () => {
       timestamp: protoTimestamp(),
     };
 
-    const result = factory.convertToInworldPacket(packet);
+    const result = EventFactory.fromProto(packet);
 
     expect(result).toBeInstanceOf(InworldPacket);
     expect(result.isEmotion()).toEqual(true);
@@ -226,7 +273,7 @@ describe('convert packet to external one', () => {
       timestamp: protoTimestamp(),
     };
 
-    const result = factory.convertToInworldPacket(packet);
+    const result = EventFactory.fromProto(packet);
 
     expect(result).toBeInstanceOf(InworldPacket);
     expect(result.isSilence()).toEqual(true);
@@ -243,7 +290,7 @@ describe('convert packet to external one', () => {
       timestamp: protoTimestamp(),
     };
 
-    const result = factory.convertToInworldPacket(packet);
+    const result = EventFactory.fromProto(packet);
 
     expect(result).toBeInstanceOf(InworldPacket);
     expect(result.isCancelResponse()).toEqual(true);
@@ -263,7 +310,7 @@ describe('convert packet to external one', () => {
       },
     };
 
-    const result = factory.convertToInworldPacket(packet);
+    const result = EventFactory.fromProto(packet);
 
     expect(result).toBeInstanceOf(InworldPacket);
     expect(result.isNarratedAction()).toEqual(true);
@@ -279,7 +326,7 @@ describe('convert packet to external one', () => {
       timestamp: protoTimestamp(),
     };
 
-    const result = factory.convertToInworldPacket(packet);
+    const result = EventFactory.fromProto(packet);
 
     expect(result).toBeInstanceOf(InworldPacket);
     expect(result.isEmotion()).toEqual(false);
@@ -304,7 +351,7 @@ describe('convert packet to external one', () => {
         timestamp: protoTimestamp(today),
       };
 
-      const result = factory.convertToInworldPacket(packet);
+      const result = EventFactory.fromProto(packet);
 
       expect(result).toBeInstanceOf(InworldPacket);
       expect(result.isControl()).toEqual(true);
@@ -325,7 +372,7 @@ describe('convert packet to external one', () => {
         timestamp: protoTimestamp(today),
       };
 
-      const result = factory.convertToInworldPacket(packet);
+      const result = EventFactory.fromProto(packet);
 
       expect(result).toBeInstanceOf(InworldPacket);
       expect(result.isControl()).toEqual(true);