+++ __tests__/services/connection.service.spec.ts
@@ -118,6 +118,18 @@ describe('history', () => {
 
     expect(clear).toHaveBeenCalledTimes(1);
   });
+
+  test('should return transcript', () => {
+    const result = 'test';
+    const getTranscript = jest
+      .spyOn(InworldHistory.prototype, 'getTranscript')
+      .mockImplementationOnce(() => result);
+
+    const transcript = connection.getTranscript();
+
+    expect(getTranscript).toHaveBeenCalledTimes(1);
+    expect(transcript).toEqual(result);
+  });
 });
 
 describe('open', () => {
@@ -436,7 +448,7 @@ describe('send', () => {
     jest
       .spyOn(GrpcAudioPlayback.prototype, 'excludeCurrentInteractionPackets')
       .mockImplementationOnce(() => [
-        eventFactory.convertToInworldPacket({
+        EventFactory.fromProto({
           ...audioEvent,
           packetId: {
             packetId: audioEvent.packetId.packetId,
@@ -515,7 +527,7 @@ describe('onMessage', () => {
     jest
       .spyOn(GrpcAudioPlayback.prototype, 'excludeCurrentInteractionPackets')
       .mockImplementationOnce(() => [
-        eventFactory.convertToInworldPacket({
+        EventFactory.fromProto({
           ...audioEvent,
           packetId: {
             ...textEvent.packetId,
@@ -571,7 +583,7 @@ describe('onMessage', () => {
     jest
       .spyOn(GrpcAudioPlayback.prototype, 'excludeCurrentInteractionPackets')
       .mockImplementationOnce(() => [
-        eventFactory.convertToInworldPacket({
+        EventFactory.fromProto({
           ...audioEvent,
           packetId: {
             ...textEvent.packetId,