+++ src/common/constants.ts
@@ -1,2 +1,3 @@
 export const GRPC_HOSTNAME = 'studio.inworld.ai';
 export const CLIENT_ID = 'web';
+export const DEFAULT_USER_NAME = 'User';