+++ src/components/history.ts
@@ -1,5 +1,7 @@
 import { v4 } from 'uuid';
 
+import { UserRequest } from '../../proto/world-engine.pb';
+import { DEFAULT_USER_NAME } from '../common/constants';
 import { Character } from '../entities/character.entity';
 import {
   Actor,
@@ -42,7 +44,6 @@ export interface HistoryItemTriggerEvent extends HistoryItemBase {
   type: CHAT_HISTORY_TYPE.TRIGGER_EVENT;
   name: string;
   outgoing?: boolean;
-  source: Actor;
 }
 
 export interface HistoryInteractionEnd extends HistoryItemBase {
@@ -52,6 +53,7 @@ export interface HistoryInteractionEnd extends HistoryItemBase {
 export interface HistoryItemNarratedAction extends HistoryItemBase {
   type: CHAT_HISTORY_TYPE.NARRATED_ACTION;
   text?: string;
+  character?: Character;
 }
 
 export type HistoryItem =
@@ -60,9 +62,14 @@ export type HistoryItem =
   | HistoryInteractionEnd
   | HistoryItemNarratedAction;
 
+interface EmotionsMap {
+  [key: string]: EmotionEvent;
+}
+
 export class InworldHistory {
   private history: HistoryItem[] = [];
   private queue: HistoryItem[] = [];
+  private emotions: EmotionsMap = {};
 
   addOrUpdate({
     characters,
@@ -75,12 +82,14 @@ export class InworldHistory {
     const utteranceId = packet.packetId?.utteranceId;
     const interactionId = packet.packetId?.interactionId;
 
-    if (packet.isText()) {
-      const id = packet.routing?.source?.isCharacter
-        ? packet.routing?.source?.name
-        : packet.routing?.target?.name;
-      const character = characters.find((x) => x.getId() === id);
+    const id = packet.routing?.source?.isCharacter
+      ? packet.routing?.source?.name
+      : packet.routing?.target?.name;
+    const character = characters.find((x) => x.getId() === id);
 
+    if (packet.isEmotion()) {
+      this.emotions[interactionId] = packet.emotions;
+    } else if (packet.isText()) {
       const actorItem: HistoryItem = {
         ...this.combineTextItem(packet),
         character,
@@ -92,7 +101,10 @@ export class InworldHistory {
         chatItem = actorItem;
       }
     } else if (packet.isNarratedAction()) {
-      const actionItem = this.combineNarratedActionItem(packet);
+      const actionItem = {
+        ...this.combineNarratedActionItem(packet),
+        character,
+      };
 
       if (
         grpcAudioPlayer.isCurrentPacket({ interactionId }) ||
@@ -199,7 +211,7 @@ export class InworldHistory {
 
         if (foundActions.length) {
           this.history = [...this.history, ...foundActions];
-          this.queue = this.queue.filter(byCondition);
+          this.queue = this.queue.filter((item) => !byCondition(item));
         }
 
         return !!foundActions.length;
@@ -226,6 +238,45 @@ export class InworldHistory {
   clear() {
     this.queue = [];
     this.history = [];
+  }
+
+  getTranscript(user?: UserRequest): string {
+    if (!this.history.length) {
+      return '';
+    }
+
+    const userName = user?.name || DEFAULT_USER_NAME;
+
+    let transcript = '';
+    let characterLastSpeaking = false;
+
+    this.history.forEach((item) => {
+      const prefix = transcript.length ? '\n' : '';
+      switch (item.type) {
+        case CHAT_HISTORY_TYPE.ACTOR:
+        case CHAT_HISTORY_TYPE.NARRATED_ACTION:
+          const isCharacter = item.source.isCharacter;
+          const givenName = isCharacter
+            ? item.character?.getDisplayName()
+            : userName;
+          const emotionCode =
+            this.emotions[item.interactionId]?.behavior?.code || '';
+          const emotion = emotionCode ? `(${emotionCode}) ` : '';
+
+          transcript +=
+            characterLastSpeaking && isCharacter
+              ? item.text
+              : `${prefix}${givenName}: ${emotion}${item.text}`;
+          characterLastSpeaking = isCharacter;
+          break;
+        case CHAT_HISTORY_TYPE.TRIGGER_EVENT:
+          transcript += `${prefix}>>> ${item.name}`;
+          characterLastSpeaking = false;
+          break;
+      }
+    });
+
+    return transcript;
   }
 
   private combineTextItem(packet: InworldPacket): HistoryItemActor {