+++ src/index.ts
@@ -23,8 +23,14 @@ import {
   HistoryItemTriggerEvent,
 } from './components/history';
 import { Character } from './entities/character.entity';
-import { EmotionBehavior } from './entities/emotion-behavior.entity';
-import { EmotionStrength } from './entities/emotion-strength.entity';
+import {
+  EmotionBehavior,
+  EmotionBehaviorCode,
+} from './entities/emotion-behavior.entity';
+import {
+  EmotionStrength,
+  EmotionStrengthCode,
+} from './entities/emotion-strength.entity';
 import {
   Actor,
   AdditionalPhonemeInfo,
@@ -36,6 +42,7 @@ import {
   Routing,
   TextEvent,
   TriggerEvent,
+  TriggerParameter,
 } from './entities/inworld_packet.entity';
 import { InworldConnectionService } from './services/inworld_connection.service';
 
@@ -50,8 +57,10 @@ export {
   ClientConfiguration,
   ConnectionConfig,
   EmotionBehavior,
+  EmotionBehaviorCode,
   EmotionEvent,
   EmotionStrength,
+  EmotionStrengthCode,
   HistoryInteractionEnd,
   HistoryItem,
   HistoryItemActor,
@@ -66,5 +75,6 @@ export {
   SessionToken,
   TextEvent,
   TriggerEvent,
+  TriggerParameter,
   User,
 };